+++ uefi-raw/src/protocol/network/ip4.rs
@@ -1,11 +1,11 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-use crate::Ipv4Address;
+use core::net::Ipv4Addr;
 
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Ord, PartialOrd, Hash)]
 #[repr(C)]
 pub struct Ip4RouteTable {
-    pub subnet_addr: Ipv4Address,
-    pub subnet_mask: Ipv4Address,
-    pub gateway_addr: Ipv4Address,
+    pub subnet_addr: Ipv4Addr,
+    pub subnet_mask: Ipv4Addr,
+    pub gateway_addr: Ipv4Addr,
 }