+++ uefi-raw/src/protocol/network/ip4_config2.rs
@@ -1,8 +1,9 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
 use crate::protocol::network::ip4::Ip4RouteTable;
-use crate::{guid, Char16, Event, Guid, Ipv4Address, MacAddress, Status};
+use crate::{guid, Char16, Event, Guid, MacAddress, Status};
 use core::ffi::c_void;
+use core::net::Ipv4Addr;
 
 newtype_enum! {
     pub enum Ip4Config2DataType: i32 => {
@@ -22,8 +23,8 @@ pub struct Ip4Config2InterfaceInfo {
     pub if_type: u8,
     pub hw_addr_size: u32,
     pub hw_addr: MacAddress,
-    pub station_addr: Ipv4Address,
-    pub subnet_mask: Ipv4Address,
+    pub station_addr: Ipv4Addr,
+    pub subnet_mask: Ipv4Addr,
     pub route_table_size: u32,
     pub route_table: *mut Ip4RouteTable,
 }
@@ -39,8 +40,8 @@ newtype_enum! {
 #[derive(Debug)]
 #[repr(C)]
 pub struct Ip4Config2ManualAddress {
-    pub address: Ipv4Address,
-    pub subnet_mask: Ipv4Address,
+    pub address: Ipv4Addr,
+    pub subnet_mask: Ipv4Addr,
 }
 
 #[derive(Debug)]