+++ xtask/src/device_path/mod.rs
@@ -79,12 +79,12 @@ fn gen_uefi_code_as_string(groups: &[NodeGroup]) -> Result<String> {
             self, DevicePathHeader, DevicePathNode, DeviceSubType, DeviceType,
             NodeConversionError,
         };
-        use crate::proto::network::IpAddress;
         use crate::mem::memory_map::MemoryType;
         use core::mem::{size_of, size_of_val};
         use core::ptr::addr_of;
         use core::{fmt, slice};
         use ptr_meta::Pointee;
+        use uefi_raw::IpAddress;
 
         #(#packed_modules)*